moved graceful killer to it's own file
