+++ gempython/tools/xdaq/set_scanParams_latency.py
@@ -8,18 +8,8 @@ from gempython.utils.gemlogger import colors,getGEMLogger,printGreen,printRed
 
 import os
 
-## https://stackoverflow.com/questions/18499497/how-to-process-sigterm-signal-gracefully
-class GracefulKiller:
-    kill_now = False
-    def __init__(self):
-        signal.signal(signal.SIGINT,  self.exit_gracefully)
-        signal.signal(signal.SIGTERM, self.exit_gracefully)
-        signal.signal(signal.SIGILL,  self.exit_gracefully)
-
-    def exit_gracefully(self,signum, frame):
-        self.kill_now = True
-
 def setScanParamsLatency(args):
+    from gempython.utils.gracefulKiller import GracefulKiller
     killer = GracefulKiller()
 
     global amc
